pyproject.toml
[tool.poetry]
name = "reeltime"
version = "0.1.0"
description = ""
authors = ["Grisha <grishamalanchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
pytest = "^7.3.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
